// src/enum.ts
export enum ConnectionCreated {
  BY_ME = "ByMe",
  BY_OTHERS = "ByOthers",
}

export enum ConnectionStatus {
  PRIVATE = "Private",
  SHARED = "Shared",
}

export enum Role {
  OWNER = "Owner",
  ADMIN = "Admin",
  EDITOR = "Editor",
  USER = "User",
  VIEWER = "Viewer",
}

export enum PermissionsText {
  OWNER = "Nekaj nekaj nekaj nekaj nekaj",
}